Steam discovery and beginning of lock logic, termination background service should be replaced by listener instead.

+++ Fenris/ProcessTerminator.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,25 +10,47 @@ namespace Fenris
 {
     public class ProcessTerminator
     {
-        public ProcessTerminator()
+        List<Process> Processes;
+        public ProcessTerminator(List<Process> processes)
         {
-            throw new NotImplementedException();
+            Processes = processes;
         }
 
-        public List<string> RetrieveProcessesToTerminate()
+        public List<Process> RetrieveProcessesToTerminate()
         {
-             throw new NotImplementedException();
+            var blockedProcesses = Processes.Where(a => a.HasBlock == true).ToList();
+            return blockedProcesses;
         }
 
-        public void TerminateProcessBackgroundService()
+        public void RunTerminateProcessBackgroundService()
         {
+            var blockedProcesses = RetrieveProcessesToTerminate();
             while (true)
             {
-                var processes = System.Diagnostics.Process.GetProcessesByName("BackgroundService");
-                foreach (var processToTerminate in processes)
+                foreach (var process in blockedProcesses)
                 {
-                    processToTerminate.Kill();
+                    if (process.BlockSettings!.IsBlockActive())
+                    {
+                        var processesToTerminate = System.Diagnostics.Process.GetProcessesByName("BackgroundService");
+                        foreach (var processToTerminate in processesToTerminate)
+                        {
+                            try
+                            {
+                                processToTerminate.Kill();
+                                Console.WriteLine($"Terminated process: {processToTerminate.ProcessName}");
+                            }
+                            catch (Exception ex)
+                            {
+                                Console.WriteLine($"Failed to terminate {processToTerminate.ProcessName}: {ex.Message}");
+                            }
+                        }
+                    }
+                    else
+                    {
+                        Console.WriteLine(process.Name + "is not blocked at the current time");
+                    }
                 }
+                Thread.Sleep(1000);
             }
         }
     }
